requirements.txt
causal_conv1d
transformers>=4.53,<4.54
torch
timm
mamba-ssm==2.2.5
accelerate
open_clip_torch
numpy
pillow
peft
trl
datasets
python-dotenv
huggingface_hub
bitsandbytes
